Extract profile storage key helper
