+++ scripts/architecture-cycle-baseline.json
@@ -1,18 +1,14 @@
 {
   "schemaVersion": 1,
-  "maxCyclicModules": 19,
-  "maxLargestCyclicComponent": 16,
+  "maxCyclicModules": 15,
+  "maxLargestCyclicComponent": 12,
   "allowedCyclicModules": [
-    "js/backup.js",
-    "js/crypto.js",
     "js/cycle-import.js",
     "js/cycle.js",
     "js/data.js",
     "js/profile.js",
     "js/sync-actions.js",
-    "js/sync-chat-apply.js",
     "js/sync-cutover.js",
-    "js/sync-payload-collectors.js",
     "js/sync-payload.js",
     "js/sync-save-hooks.js",
     "js/sync-storage-cleanup.js",
