+++ tsconfig.checkjs.json
@@ -213,6 +213,7 @@
     "js/pii.js",
     "js/profile-fatty-acid-migrations.js",
     "js/profile-marker-migrations.js",
+    "js/profile-storage-key.js",
     "js/profile-share.js",
     "js/profile.js",
     "js/profile-runtime.js",
